==> notebooks/00_exploration.ipynb <==
import pandas as pd 
import numpy as np
pd.options.display.max_rows = 10000

DATA_PATH = '../data/raw/survey_results_public.csv'

# Read data and print shape
raw_df = pd.read_csv(DATA_PATH)
raw_df.shape

# Display random answer 
# Observations: Multiple answers need to be splitted 
# Reference to the schema needed to understand
raw_df.sample(1).iloc[0]

# Print the general information of the data frame 
raw_df.info()

# Get stats for the numerical column
raw_df.describe()

# get summarize Categorical data also
raw_df[['LanguageHaveWorkedWith',
                  'DatabaseHaveWorkedWith',
                  'PlatformHaveWorkedWith',
                  'WebframeHaveWorkedWith',
                  'MiscTechHaveWorkedWith',
                  'ToolsTechHaveWorkedWith',
                  'NEWCollabToolsHaveWorkedWith',"Currency", "DevType"]].describe(include = 'object')

# Investigate the questionable objects columns
questionable_cols = ['YearsCodePro', 'YearsCode']

for col in questionable_cols: 
    print(col)
    print(raw_df[col].unique().tolist())
    print('--------------------------')
    print()
